requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.17.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.2.0
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
executing==2.2.1
filelock==3.20.0
flake8==7.3.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hydra-core==1.3.2
icecream==2.1.8
idna==3.11
ImageIO==2.37.2
ipykernel==7.0.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mccabe==0.7.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
optuna==4.5.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
polars==1.34.0
polars-runtime-32==1.34.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycodestyle==2.14.0
pycodestyle_magic==0.5
pydantic==2.12.0
pydantic_core==2.41.1
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
pystache==0.6.8
python-dateutil==2.9.0.post0
pytorch-triton-rocm==3.4.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
sentry-sdk==2.41.0
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
stringzilla==4.2.1
sympy==1.14.0
termcolor==3.1.0
tifffile==2025.10.16
timm==1.0.20
torch==2.8.0+rocm6.3
torchaudio==2.8.0+rocm6.3
torchinfo==1.8.0
torchmetrics==1.8.2
torchview==0.2.7
torchvision==0.23.0+rocm6.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.25.1
wcwidth==0.2.14
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
